Make Generator class
